Make the "-c" option for the build script do what I expected it to do.
Now it cleans up and rebuilds. Previously it just did a 'make clean'. If
the previous build was 'build -p', this meant that the public copies of
the *.so files would be deleted. Oddly, the users didn't seem to like
that.

To get the previous behaviour, use the option twice: 'build -c -c'.