Check for version 1 or 2 of VisIt before building
Only major version 1 or 2 of VisIt are supported by the build
script. Anything else should be flagged as an error.
This change also detects if visit has been left in its build
location which does not support the addition of 3rd party plugins.

This fixes SDF/SDF#4 and also fixes EPOCH/epoch#1320